Cleanup rename project OpenSearch to SearchEngine and add folders for each engine
