+++ src/dms/Dockerfile
@@ -14,7 +14,7 @@ COPY backend/EdFi.DataManagementService.Backend.Installer/ backend/EdFi.DataMana
 COPY backend/EdFi.DataManagementService.Backend/ backend/EdFi.DataManagementService.Backend/
 COPY backend/EdFi.DataManagementService.Backend.Mssql/ backend/EdFi.DataManagementService.Backend.Mssql/
 COPY backend/EdFi.DataManagementService.Backend.Postgresql/ backend/EdFi.DataManagementService.Backend.Postgresql/
-COPY backend/EdFi.DataManagementService.Backend.OpenSearch/ backend/EdFi.DataManagementService.Backend.OpenSearch/
+COPY backend/EdFi.DataManagementService.Backend.SearchEngine/ backend/EdFi.DataManagementService.Backend.SearchEngine/
 COPY --from=parentdir .editorconfig .
 COPY --from=parentdir Directory.Packages.props ./
 COPY --from=parentdir nuget.config .
