+++ src/dms/backend/EdFi.DataManagementService.Backend.SearchEngine/OpenSearch/QueryOpenSearch.cs
@@ -12,7 +12,7 @@ using Microsoft.Extensions.Logging;
 using OpenSearch.Client;
 using OpenSearch.Net;
 
-namespace EdFi.DataManagementService.Backend.OpenSearch;
+namespace EdFi.DataManagementService.Backend.SearchEngine;
 
 /// <summary>
 /// Queries OpenSearch for documents. Example query DSL usage:
