+++ src/dms/frontend/EdFi.DataManagementService.Frontend.AspNetCore/Infrastructure/WebApplicationBuilderExtensions.cs
@@ -8,7 +8,8 @@ using System.Security.Claims;
 using System.Threading.RateLimiting;
 using EdFi.DataManagementService.Backend;
 using EdFi.DataManagementService.Backend.Deploy;
-using EdFi.DataManagementService.Backend.OpenSearch;
+using EdFi.DataManagementService.Backend.SearchEngine.elasticsearch;
+using EdFi.DataManagementService.Backend.SearchEngine.OpenSearch;
 using EdFi.DataManagementService.Backend.Postgresql;
 using EdFi.DataManagementService.Core;
 using EdFi.DataManagementService.Core.OAuth;
